scripts/create_summary_poster.py
"""
Create a comprehensive summary poster combining all key results.
Perfect for presentations or as a graphical abstract.
"""

import matplotlib.pyplot as plt
import matplotlib.patches as patches
from matplotlib.gridspec import GridSpec
import numpy as np
from pathlib import Path
from PIL import Image

def create_summary_poster(figures_dir, output_path):
    """Create a comprehensive summary poster."""
    
    figures_dir = Path(figures_dir)
    
    # Create large figure
    fig = plt.figure(figsize=(20, 14))
    gs = GridSpec(4, 3, figure=fig, hspace=0.3, wspace=0.3)
    
    # Title
    fig.suptitle('Hybrid Offline RL for Robotic Pick-and-Place: 96.67% Success Rate',
                 fontsize=24, fontweight='bold', y=0.98)
    
    # Row 1: Task Overview
    ax_task = fig.add_subplot(gs[0, :])
    ax_task.text(0.5, 0.8, 'Task: 7-DoF Robotic Arm Pick-and-Place', 
                ha='center', va='top', fontsize=18, fontweight='bold')
    ax_task.text(0.5, 0.5, 
                '6 Phases: APPROACH → DESCEND → GRASP → LIFT → MOVE → FINE',
                ha='center', va='center', fontsize=14, style='italic')
    ax_task.text(0.5, 0.2,
                'Dataset: 38,065 samples | 813 episodes | Offline learning',
                ha='center', va='center', fontsize=12)
    ax_task.axis('off')
    ax_task.set_xlim(0, 1)
    ax_task.set_ylim(0, 1)
    
    # Row 2: Load and display key figures
    try:
        # Environment
        ax_env = fig.add_subplot(gs[1, 0])
        img_env = Image.open(figures_dir / 'environment_screenshots.png')
        ax_env.imshow(img_env)
        ax_env.set_title('Environment', fontsize=14, fontweight='bold')
        ax_env.axis('off')
    except:
        ax_env = fig.add_subplot(gs[1, 0])
        ax_env.text(0.5, 0.5, 'Environment\nScreenshots', ha='center', va='center', fontsize=12)
        ax_env.axis('off')
    
    try:
        # Architecture
        ax_arch = fig.add_subplot(gs[1, 1])
        img_arch = Image.open(figures_dir / 'architecture_diagram.png')
        ax_arch.imshow(img_arch)
        ax_arch.set_title('System Architecture', fontsize=14, fontweight='bold')
        ax_arch.axis('off')
    except:
        ax_arch = fig.add_subplot(gs[1, 1])
        ax_arch.text(0.5, 0.5, 'Architecture\nDiagram', ha='center', va='center', fontsize=12)
        ax_arch.axis('off')
    
    try:
        # Phase distribution
        ax_phase = fig.add_subplot(gs[1, 2])
        img_phase = Image.open(figures_dir / 'phase_distribution.png')
        ax_phase.imshow(img_phase)
        ax_phase.set_title('Dataset Distribution', fontsize=14, fontweight='bold')
        ax_phase.axis('off')
    except:
        ax_phase = fig.add_subplot(gs[1, 2])
        ax_phase.text(0.5, 0.5, 'Phase\nDistribution', ha='center', va='center', fontsize=12)
        ax_phase.axis('off')
    
    # Row 3: Main Results
    ax_results = fig.add_subplot(gs[2, :])
    
    # Create bar chart
    methods = ['BC\nBaseline', 'BC + IQL\nCritic\n(Ours)', 'Diffusion\n(DDIM)', 'Diffusion\n(Critic-Guided)']
    success_rates = [85.0, 96.67, 35.0, 30.0]
    colors = ['#3498db', '#2ecc71', '#e74c3c', '#f39c12']
    
    bars = ax_results.bar(methods, success_rates, color=colors, alpha=0.8, 
                          edgecolor='black', linewidth=3)
    
    # Highlight best
    bars[1].set_edgecolor('gold')
    bars[1].set_linewidth(5)
    
    # Add value labels
    for bar, rate in zip(bars, success_rates):
        height = bar.get_height()
        ax_results.text(bar.get_x() + bar.get_width()/2., height,
                       f'{rate:.1f}%', ha='center', va='bottom', 
                       fontsize=16, fontweight='bold')
    
    ax_results.axhline(y=90, color='green', linestyle='--', linewidth=2, alpha=0.5)
    ax_results.set_ylabel('Success Rate (%)', fontsize=14, fontweight='bold')
    ax_results.set_title('MAIN RESULTS: Performance Comparison', fontsize=16, fontweight='bold')
    ax_results.set_ylim([0, 105])
    ax_results.grid(axis='y', alpha=0.3)
    
    # Row 4: Key Contributions
    ax_contrib = fig.add_subplot(gs[3, :])
    
    contributions = [
        '✓ Phase-Based Task Decomposition',
        '✓ Balanced Sampling Strategy',
        '✓ Critic-Guided Action Selection',
        '✓ Twin Q-Networks for Stability',
        '✓ Adaptive Candidate Generation',
        '✓ 96.67% Success Rate (SOTA)'
    ]
    
    y_pos = 0.9
    for contrib in contributions:
        ax_contrib.text(0.05, y_pos, contrib, fontsize=13, fontweight='bold',
                       verticalalignment='top')
        y_pos -= 0.15
    
    # Add summary box
    summary_text = (
        'SUMMARY:\n'
        'Hybrid BC-IQL approach achieves 96.67% success rate,\n'
        'outperforming BC baseline (+11.67%) and diffusion policies (+61.67%).\n'
        'Training: ~30 min | Inference: Real-time | Dataset: 38K samples'
    )
    
    ax_contrib.text(0.98, 0.5, summary_text, fontsize=11,
                   verticalalignment='center', horizontalalignment='right',
                   bbox=dict(boxstyle='round', facecolor='wheat', alpha=0.8, pad=1))
    
    ax_contrib.set_xlim(0, 1)
    ax_contrib.set_ylim(0, 1)
    ax_contrib.axis('off')
    
    # Save
    plt.savefig(output_path, dpi=300, bbox_inches='tight', facecolor='white')
    plt.close()
    print(f"✓ Summary poster saved: {output_path}")


if __name__ == "__main__":
    import argparse
    parser = argparse.ArgumentParser()
    parser.add_argument("--figures_dir", type=str, default="figures")
    parser.add_argument("--output", type=str, default="figures/SUMMARY_POSTER.png")
    args = parser.parse_args()
    
    create_summary_poster(args.figures_dir, args.output)
