requirements.txt
mujoco==3.1.6
gymnasium==0.29.1
numpy>=1.26.0
torch==2.3.1
tqdm>=4.66.0
opencv-python>=4.9.0
Pillow>=10.0.0
pydantic>=2.6.0
huggingface-hub>=0.22.0
matplotlib>=3.7.0
seaborn>=0.12.0
